backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pydantic
python-dotenv
httpx
